io: port over optimisations and other changes from mio2it
Most notably setError not being inlined reduces a lot of exception
throwing related code bloat from functions that use it.
Exception throwing shouldn't be optimised in that way because they
should not happen often. Surely in our case it being inlined is slower
because it's making the function larger just to fit a really cold path,
which would mean less would fit in cache?

Also in here is that the FileIO open functions are now public, because
before `close` was but `open` wasn't??? Perhaps similar changes should
be made in DynBufIO and SpanIO too (swap out which buffers to use)?