mpb: support 1.50
MPB 1.50 is used in Sega Rally 2. It seems the panning calculation
matches that of version 2. Thanks to Pitito for letting me know.

Additionally, this makes the MPB tools show the version as hexadecimal
instead of denary. Ideally this would be in the other tools too but I'm
poorly and want to get off a computer

(Additionally, when versions should be printed, it should really be
split into a char[4] with a period delimiter, i.e, "1.50" as opposed to
5001. Should file an issue to remind myself really)